+++ android/build.gradle
@@ -1,18 +1,9 @@
-buildscript {
-    ext.kotlin_version = '2.0.20' // Update to the latest stable Kotlin version
-
+allprojects {
     repositories {
         google()
         mavenCentral()
-        jcenter()
-    }
-
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.android.tools.build:gradle:2.3.0'
     }
 }
-
 
 rootProject.buildDir = "../build"
 subprojects {
