+++ lib/main.dart
@@ -16,4 +16,4 @@ class MainApp extends StatelessWidget {
       home: HomeView(),
     );
   }
-
+}
