--- lib/ui/utils/auth_validators.dart
@@ -1,24 +1,0 @@
-class AuthValidators {
-  static String? validateEmail(String? email) {
-    if (email == null || email.isEmpty) {
-      return 'E-Posta adresinizi girin';
-    }
-    final emailRegex = RegExp(
-      r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$',
-    );
-    if (!emailRegex.hasMatch(email)) {
-      return 'Geçerli bir e-posta adresi girin';
-    }
-    return null;
-  }
-
-  static String? validatePassword(String? password) {
-    if (password == null || password.isEmpty) {
-      return 'Şifrenizi girin';
-    }
-    if (password.length < 6) {
-      return 'Şifreniz en az 6 karakter olmalıdır';
-    }
-    return null;
-  }
-}
