+++ pubspec.yaml
@@ -7,20 +7,13 @@ environment:
   sdk: ^3.5.1
 
 dependencies:
-  cached_network_image: ^3.4.1
   flutter:
     sdk: flutter
-  webview_flutter: ^4.10.0
-
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^4.0.0
-  build_runner:
-  flutter_gen_runner:
 
 flutter:
   uses-material-design: true
-  assets:
-        - assets/images/
